add clear agent service rule
