+++ apm-sniffer/apm-agent-core/src/test/java/org/skywalking/apm/agent/core/boot/ServiceManagerTest.java
@@ -3,7 +3,10 @@ package org.skywalking.apm.agent.core.boot;
 import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.List;
+import org.junit.Rule;
 import org.junit.Test;
+import org.skywalking.apm.agent.core.boot.BootService;
+import org.skywalking.apm.agent.core.boot.ServiceManager;
 import org.skywalking.apm.agent.core.context.ContextManager;
 import org.skywalking.apm.agent.core.context.IgnoredTracerContext;
 import org.skywalking.apm.agent.core.context.TracingContext;
@@ -14,6 +17,7 @@ import org.skywalking.apm.agent.core.remote.GRPCChannelListener;
 import org.skywalking.apm.agent.core.remote.GRPCChannelManager;
 import org.skywalking.apm.agent.core.remote.TraceSegmentServiceClient;
 import org.skywalking.apm.agent.core.sampling.SamplingService;
+import org.skywalking.apm.agent.core.test.tools.AgentServiceRule;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -21,9 +25,12 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 public class ServiceManagerTest {
+
+    @Rule
+    public AgentServiceRule agentServiceRule = new AgentServiceRule();
+
     @Test
     public void testServiceDependencies() throws Exception {
-        ServiceManager.INSTANCE.boot();
         HashMap<Class, BootService> registryService = getFieldValue(ServiceManager.INSTANCE, "bootedServices");
 
         assertThat(registryService.size(), is(7));
