+++ apm-sniffer/apm-agent-core/src/test/java/org/skywalking/apm/agent/core/test/tools/SegmentStoragePoint.java
@@ -1,4 +1,4 @@
-package org.skywalking.apm.agent.core.context.util;
+package org.skywalking.apm.agent.core.test.tools;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
